+++ source/pcraster_dal/dal_GSLIBBlockDriver.cc
@@ -138,7 +138,7 @@ void GSLIBBlockDriver::write(
   size_t i = 0;
 
   // Store pointers to stacks for efficiency.
-  for(int row = block.nrRows() - 1; row >= 0; --row) {
+  for(long row = block.nrRows() - 1; row >= 0; --row) {
     for(size_t col = 0; col < block.nrCols(); ++col) {
       i = row * block.nrCols() + col;
       stacks[i] = &block.cell<std::vector<T> >(i);
@@ -146,7 +146,7 @@ void GSLIBBlockDriver::write(
   }
 
   for(size_t voxel = 0; voxel < nrVoxelsPerStack; ++voxel) {
-    for(int row = block.nrRows() - 1; row >= 0; --row) {
+    for(long row = block.nrRows() - 1; row >= 0; --row) {
       for(size_t col = 0; col < block.nrCols(); ++col) {
         i = row * block.nrCols() + col;
         if(pcr::isMV((*stacks[i])[voxel])) {
@@ -174,7 +174,7 @@ void GSLIBBlockDriver::write<UINT1>(
   size_t i = 0;
 
   // Store pointers to stacks for efficiency.
-  for(int row = block.nrRows() - 1; row >= 0; --row) {
+  for(long row = block.nrRows() - 1; row >= 0; --row) {
     for(size_t col = 0; col < block.nrCols(); ++col) {
       i = row * block.nrCols() + col;
       stacks[i] = &block.cell<std::vector<UINT1> >(i);
@@ -182,7 +182,7 @@ void GSLIBBlockDriver::write<UINT1>(
   }
 
   for(size_t voxel = 0; voxel < nrVoxelsPerStack; ++voxel) {
-    for(int row = block.nrRows() - 1; row >= 0; --row) {
+    for(long row = block.nrRows() - 1; row >= 0; --row) {
       for(size_t col = 0; col < block.nrCols(); ++col) {
         i = row * block.nrCols() + col;
         if(pcr::isMV((*stacks[i])[voxel])) {
